requirements.txt
absl-py==1.4.0
albumentations==1.4.20
appdirs==1.4.4
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.4
cmake==3.26.4
contourpy==1.1.0
cycler==0.11.0
debugpy==1.5.1
decord==0.6.0
docker-pycreds==0.4.0
filelock==3.12.2
fonttools==4.40.0
gitdb==4.0.10
gitpython==3.1.31
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.56.2
idna==3.4
imageio==2.37.0
jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kornia==0.6.12
lazy-loader==0.4
lit==16.0.6
markdown==3.4.3
markupsafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
natsort==8.4.0
networkx==3.1
numpy==1.25.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
oauthlib==3.2.2
opencv-python==4.7.0.72
pandas==2.0.3
pathtools==0.1.2
pillow==11.3.0
protobuf==4.23.4
psutil==5.9.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.8
pyparsing==3.1.0
pytz==2023.3
pyyaml==6.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.16.1
sentry-sdk==1.27.1
setproctitle==1.3.2
setuptools==75.8.2
smmap==5.0.0
swanlab==0.6.8
sympy==1.13.3
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorboardx==2.6.2.2
threadpoolctl==3.1.0
tifffile==2023.9.26
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tornado==6.1
tqdm==4.65.0
triton==3.3.1
tzdata==2023.3
urllib3==1.26.16
wandb==0.15.5
werkzeug==2.3.6
wheel==0.45.1